+++ hirpc-kit/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this package will be documented in this file.
 
+## [0.1.5] - 2025-06-10
+
+### Added
+
+- Support for LAZY_HIRPC_LOAD and hiRPC v0.6.0
+
 ## [0.1.4] - 2025-05-30
 
 ### Added