+++ unity/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this package will be documented in this file.
 
+## [2.1.0] - 2025-06-10
+
+## Added
+
+- LazyHiRpcLoad option in the Advanced Configuration
+- Api.Commands.GetRelationships (requires Discord approval)
+- Api.Subscribe.RelationshipUpdate (requires Discord approval)
+- Models.Relationship
+- Relationships section in the Discord Mock
+
+## Changed
+
+- Minor performance improvements related to the hiRPC Interface
+
+## Fixed
+
+- The game doesn't turn into a black screen when loading the index.html
+
 ## [2.0.1] - 2025-05-30
 
 ### Added